requirements.txt
pandas==2.2.2
fsspec==2024.3.1          # permite leer hf://
torch==2.2.2          # pip install --index-url https://download.pytorch.org/whl/cpu
torchtext==0.17.2
transformers==4.40.1
datasets==2.19.0
evaluate==0.4.2
optuna==3.6.1
scikit-learn==1.4.2
seaborn==0.13.2
matplotlib==3.8.4
plotly==5.21.0
wordcloud==1.9.3
streamlit==1.35.0
